Identify the answer to s1c3
https://cryptopals.com/sets/1/challenges/3

By running:

python -m set1.challenge3_1_byte_XOR_cipher.s1c3

and inspecting the output showing the top scorers on my current metric,
which is vector distance to letter frequencies, I identified the answer
as "X" (repeated so that its hex encoding is as long as the hex-encoded
ciphertext) for the key, and "Cooking MC's like a pound of bacon" for
the encrypted message.

My metric did not score the correct key with the lowest score. It was in
the top 7 or so, but this suggests my metric could be revised to
consider:
  - capitals versus non-capitals
  - spaces, word length, etc
  - non-letter characters